+++ src/renderer/components/screens/AccountListScreen.tsx
@@ -261,7 +261,7 @@ function AccountItem({
 
   const updateUnreadCount = useMemo(
     () =>
-      debounce((_ev: any, account_id: number) => {
+      debounce((account_id: number) => {
         if (account_id === login.id) {
           BackendRemote.rpc
             .getFreshMsgs(login.id)
@@ -273,14 +273,13 @@ function AccountItem({
   )
 
   useEffect(() => {
-    updateUnreadCount(null, login.id)
+    updateUnreadCount(login.id)
     // TODO use onIncomingMsg event directly after we changed the events to be filtered for active account in the frontend
-    ipcBackend.on('DD_EVENT_INCOMING_MESSAGE_ACCOUNT', updateUnreadCount)
+    let emitter = BackendRemote.getContextEvents(login.id)
+    
+    emitter.on("IncomingMsg", updateUnreadCount.bind(null, login.id))
     return () => {
-      ipcBackend.removeListener(
-        'DD_EVENT_INCOMING_MESSAGE_ACCOUNT',
-        updateUnreadCount
-      )
+      emitter.off("IncomingMsg", updateUnreadCount.bind(null, login.id))
     }
   }, [login.id, updateUnreadCount])